get prod_list

+++ migrations/items.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS public.items
 (
 id integer PRIMARY KEY,
- product_id integer NOT NULL,
+ product_id bigint NOT NULL,
   offer_id varchar NOT NULL,
    is_fbo_visible bool NOT NULL,
     is_fbs_visible bool NOT NULL,
